requirements.txt
astunparse==1.6.3
certifi==2021.10.8
jittor==1.3.3.14
numpy==1.21.6
opencv-python==4.5.5.64
Pillow==9.1.0
protobuf==3.20.1
scipy==1.7.3
six==1.16.0
tensorboardX==2.5
tqdm==4.64.0
